// src/app/interfaces/mainProduct.ts
export interface MainProduct {
    _id: string;
    name: string;
    description: string;
    price: number;
    imageUrl: string;
    initial_price: number;
    quantity:number;
    packing:string;
    hsnocde:number;
    cgst:number;
    sgst:number;
    igst:number;
    status: string;
    company_name:string;
}